Fix infinite loop when case folding invalid UTF8 chars

+++ src/utf8.h
@@ -6,8 +6,8 @@
 
 void utf8proc_case_fold(strbuf *dest, const uint8_t *str, int len);
 void utf8proc_encode_char(int32_t uc, strbuf *buf);
-ssize_t utf8proc_iterate(const uint8_t *str, ssize_t str_len, int32_t *dst);
-ssize_t utf8proc_charlen(const uint8_t *str, ssize_t str_len);
+int utf8proc_iterate(const uint8_t *str, int str_len, int32_t *dst);
+int utf8proc_charlen(const uint8_t *str, int str_len);
 void utf8proc_detab(strbuf *dest, const uint8_t *line, size_t size);
 
 #endif
